put backward compat code in a separate module
