+++ src/game/mod.rs
@@ -13,6 +13,7 @@ use tombstone::Tombstone;
 
 pub mod battle;
 pub mod tombstone;
+mod game040;
 
 #[derive(Debug)]
 pub enum Error {
@@ -49,7 +50,7 @@ impl Game {
         } else {
             // if json deserialization fails, attempt bincode assuming
             // it may be a file from v0.4.0
-            Game040::deserialize(&data).unwrap()
+            game040::deserialize(&data).unwrap()
         };
         Ok(game)
     }
@@ -223,36 +224,6 @@ impl Game {
 impl Default for Game {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-/// v0.4.0 of the game struct, kept for backwards compatibility when upgrading
-/// the game data file to the new version
-// FIXME remove on a subsequent version
-#[derive(Deserialize)]
-struct Game040 {
-    pub player: Character,
-    pub location: Location,
-    pub gold: i32,
-    inventory: HashMap<String, Vec<Box<dyn Item>>>,
-    tombstones: HashMap<Location, Tombstone>,
-}
-
-impl Game040 {
-    /// Get a new Game instance out of a v0.4.0 one
-    fn deserialize(data: &[u8]) -> Result<Game, bincode::Error> {
-        let mut v4game: Game040 = bincode::deserialize(&data)?;
-        let mut new_game = Game::new();
-        std::mem::swap(&mut new_game.player, &mut v4game.player);
-        std::mem::swap(&mut new_game.location, &mut v4game.location);
-        std::mem::swap(&mut new_game.inventory, &mut v4game.inventory);
-        new_game.tombstones = v4game
-            .tombstones
-            .drain()
-            .map(|(l, t)| (l.to_string(), t))
-            .collect();
-        new_game.gold = v4game.gold;
-        Ok(new_game)
     }
 }
